requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
pycodestyle
google-genai
python-dotenv